@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using NIST.CVP.Common.Config;
 using NIST.CVP.Common.Interfaces;
 using NIST.CVP.Common.Oracle;
@@ -44,7 +45,8 @@ namespace NIST.CVP.PoolAPI
                     });
             });
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddControllersWithViews()
+                .AddNewtonsoftJson();
 
             services.AddSingleton(Configuration);
 
@@ -65,7 +67,7 @@ namespace NIST.CVP.PoolAPI
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IHostEnvironment env)
         {
             LogManager.GetCurrentClassLogger().Info("Configuring Startup service...");
 
@@ -80,8 +82,11 @@ namespace NIST.CVP.PoolAPI
                 //app.UseHsts();
             }
 
-            //app.UseHttpsRedirection();
-            app.UseMvc();
+            app.UseRouting();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
 
             LogManager.GetCurrentClassLogger().Info("Startup service configured.");
         }
