@@ -8,8 +8,8 @@
         /// <summary>
         /// Generate <see cref="ITestVectorSet"/> based on an algorithms registration json
         /// </summary>
-        /// <param name="requestFilePath">The full file path to the registration json</param>
+        /// <param name="generateRequest">The json representing an <see cref="IParameters"/>.</param>
         /// <returns></returns>
-        GenerateResponse Generate(string requestFilePath);
+        GenerateResponse Generate(GenerateRequest generateRequest);
     }
 }
