@@ -3,6 +3,6 @@ namespace NIST.CVP.Generation.Core
 {
     public interface IParameterChecker
     {
-        ParameterCheckResponse CheckParameters(string registrationFile);
+        ParameterCheckResponse CheckParameters(ParameterCheckRequest request);
     }
 }
