@@ -28,39 +28,35 @@ namespace NIST.CVP.Generation.Core.Tests
         public virtual string Revision { get; set; } = "1.0";
         public virtual IJsonConverterProvider JsonConverterProvider => new JsonConverterProvider();
 
-        public string DllDropLocation { get; private set; }
-
-        public string TestPath { get; private set; }
-        public string JsonSavePath { get; private set; }
+        private string TestPath { get; set; }
+        private string JsonSavePath { get; set; }
 
         public IRegisterInjections RegistrationsOracle => new RegisterInjections();
         public abstract IRegisterInjections RegistrationsGenVal { get; }
 
-        public string[] TestVectorFileNames = { "expectedResults.json", "internalProjection.json", "prompt.json" };
+        private readonly string[] _testVectorFileNames = { "expectedResults.json", "internalProjection.json", "prompt.json" };
 
         protected abstract void ModifyTestCaseToFail(dynamic testCase);
         protected abstract string GetTestFileFewTestCases(string folderName);
         protected abstract string GetTestFileLotsOfTestCases(string folderName);
 
         // Set this during a test if you want to save the json from the session
-        public bool SaveJson = true;
+        private bool SaveJson = true;
 
-        protected static Logger GenLogger => LogManager.GetLogger("Generator");
-        protected static Logger ValLogger => LogManager.GetLogger("Validator");
+        private static readonly Logger GenLogger = LogManager.GetLogger("Generator");
+        private static readonly Logger ValLogger = LogManager.GetLogger("Validator");
 
-        protected static readonly string RootDirectory = AppDomain.CurrentDomain.BaseDirectory;
-        protected IConfigurationRoot ConfigurationRoot;
-        protected IServiceProvider ServiceProvider;
+        private static readonly string RootDirectory = AppDomain.CurrentDomain.BaseDirectory;
+        private IConfigurationRoot ConfigurationRoot;
+        private IServiceProvider ServiceProvider;
 
+        private static readonly IFileService FileService = new FileService();
+        
         [OneTimeSetUp]
         public virtual void OneTimeSetUp()
         {
             TestPath = Utilities.GetConsistentTestingStartPath(GetType(), @"..\..\TestFiles\temp_integrationTests\");
             JsonSavePath = Utilities.GetConsistentTestingStartPath(GetType(), @"..\..\..\..\json-files\");
-
-            DllDropLocation =
-                Utilities.GetConsistentTestingStartPath(GetType(),
-                    @"..\..\..\common\src\NIST.CVP.Generation.GenValApp\");
 
             ConfigurationRoot = EntryPointConfigHelper.GetConfigurationRoot(RootDirectory);
             var serviceCollection = EntryPointConfigHelper.GetBaseServiceCollection(ConfigurationRoot);
@@ -125,9 +121,9 @@ namespace NIST.CVP.Generation.Core.Tests
                 Directory.CreateDirectory(JsonSavePath);
             }
 
-            var spacing = Mode == "" ? "" : "-";
+            var modeSpacing = Mode == "" ? "" : "-";
             var friendlyAlgorithm = Algorithm.Replace("/", "-");
-            var newLocation = Path.Combine(JsonSavePath, $"{friendlyAlgorithm}{spacing}{Mode}{spacing}{Revision}");
+            var newLocation = Path.Combine(JsonSavePath, $"{friendlyAlgorithm}{modeSpacing}{Mode}-{Revision}");
             if (Directory.Exists(newLocation))
             {
                 Directory.Delete(newLocation, true);
@@ -213,35 +209,37 @@ namespace NIST.CVP.Generation.Core.Tests
 
         protected void RunGeneration(string targetFolder, string fileName, bool overrideRegisteredDependencies)
         {
+            var registrationJson = FileService.ReadFile(fileName);
+            
             // Run test vector generation
             using (var scope = GetContainer(overrideRegisteredDependencies).BeginLifetimeScope())
             {
                 var gen = scope.Resolve<IGenerator>();
-                var result = gen.Generate(fileName);
+                var result = gen.Generate(new GenerateRequest(registrationJson));
 
+                FileService.WriteFile(Path.Combine(targetFolder, _testVectorFileNames[0]), result.ResultProjection, true);
+                FileService.WriteFile(Path.Combine(targetFolder, _testVectorFileNames[1]), result.InternalProjection, true);
+                FileService.WriteFile(Path.Combine(targetFolder, _testVectorFileNames[2]), result.PromptProjection, true);
+                
                 Assert.IsTrue(result.Success, $"Generator failed to complete with status code: {result.StatusCode}, {EnumHelpers.GetEnumDescriptionFromEnum(result.StatusCode)}, {result.ErrorMessage}");
             }
-
-            Assert.IsTrue(File.Exists(Path.Combine(targetFolder, TestVectorFileNames[0])), Path.Combine(targetFolder, TestVectorFileNames[0]));
-            Assert.IsTrue(File.Exists(Path.Combine(targetFolder, TestVectorFileNames[1])), Path.Combine(targetFolder, TestVectorFileNames[1]));
-            Assert.IsTrue(File.Exists(Path.Combine(targetFolder, TestVectorFileNames[2])), Path.Combine(targetFolder, TestVectorFileNames[2]));
         }
 
         protected void RunValidation(string targetFolder)
         {
+            var resultJson = FileService.ReadFile(Path.Combine(targetFolder, _testVectorFileNames[0]));
+            var internalJson = FileService.ReadFile(Path.Combine(targetFolder, _testVectorFileNames[1]));
+            
             // Run test vector validation
             using (var scope = GetContainer().BeginLifetimeScope())
             {
                 var val = scope.Resolve<IValidator>();
-                var result = val.Validate(
-                    Path.Combine(targetFolder, TestVectorFileNames[0]),
-                    Path.Combine(targetFolder, TestVectorFileNames[1]),
-                    showExpected: true
-                );
+                var result = val.Validate(new ValidateRequest(internalJson, resultJson, true));
 
+                FileService.WriteFile(Path.Combine(targetFolder, "validation.json"), result.ValidationResult, true);
+                
                 Assert.IsTrue(result.Success, $"Validator failed to complete with status code: {result.StatusCode}, {EnumHelpers.GetEnumDescriptionFromEnum(result.StatusCode)}, {result.ErrorMessage}");
             }
-            Assert.IsTrue(File.Exists(Path.Combine(targetFolder, "validation.json")), $"{targetFolder} validation");
         }
 
         private IContainer GetContainer(bool overrideRegisteredDependencies = false)
@@ -268,7 +266,7 @@ namespace NIST.CVP.Generation.Core.Tests
 
         private void GetFailureTestCases(string targetFolder, ref List<int> failureTcIds)
         {
-            var files = GetFileNamesWithPath(targetFolder, TestVectorFileNames);
+            var files = GetFileNamesWithPath(targetFolder, _testVectorFileNames);
 
             // Modify testResults in order to contain some tests that will fail
             var expectedFailTestCases = DoBadThingsToResultsFile(files[0]);
