@@ -19,8 +19,6 @@ namespace NIST.CVP.Generation.Core.Tests
     [TestFixture, UnitTest]
     public class GeneratorTests
     {
-        private string _testPath;
-        
         private Mock<ITestVectorFactory<FakeParameters, FakeTestVectorSet, FakeTestGroup, FakeTestCase>> _mockITestVectorFactory;
         private Mock<ITestCaseGeneratorFactoryFactory<FakeTestVectorSet, FakeTestGroup, FakeTestCase>> _mockITestCaseGeneratorFactoryFactory;
         private Mock<IParameterParser<FakeParameters>> _mockIParameterParser;
@@ -70,19 +68,6 @@ namespace NIST.CVP.Generation.Core.Tests
             );
         }
 
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            _testPath = Utilities.GetConsistentTestingStartPath(GetType(), "../../TestFiles/GeneratorTests/");
-            Directory.CreateDirectory(_testPath);
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            Directory.Delete(_testPath, true);
-        }
-
         [Test]
         public void GenerateShouldReturnErrorResponseWhenParametersNotParsedSuccessfully()
         {
@@ -91,7 +76,7 @@ namespace NIST.CVP.Generation.Core.Tests
                 .Setup(s => s.Parse(It.IsAny<string>()))
                 .Returns(() => new ParseResponse<FakeParameters>(errorMessage));
 
-            var result = _subject.Generate(string.Empty);
+            var result = _subject.Generate(new GenerateRequest(string.Empty));
 
             Assert.IsFalse(result.Success);
             Assert.AreEqual(errorMessage, result.ErrorMessage);
@@ -106,7 +91,7 @@ namespace NIST.CVP.Generation.Core.Tests
                 .Setup(s => s.Validate(It.IsAny<FakeParameters>()))
                 .Returns(() => new ParameterValidateResponse(errorMessage));
 
-            var result = _subject.Generate(string.Empty);
+            var result = _subject.Generate(new GenerateRequest(string.Empty));
 
             Assert.IsFalse(result.Success);
             Assert.AreEqual(errorMessage.First(), result.ErrorMessage);
@@ -121,7 +106,7 @@ namespace NIST.CVP.Generation.Core.Tests
                 .Setup(s => s.BuildTestCases(It.IsAny<FakeTestVectorSet>()))
                 .Returns(() => new GenerateResponse(errorMessage, StatusCode.TestCaseGeneratorError));
 
-            var result = _subject.Generate(string.Empty);
+            var result = _subject.Generate(new GenerateRequest(string.Empty));
 
             Assert.IsFalse(result.Success);
             Assert.AreEqual(errorMessage, result.ErrorMessage);
@@ -133,58 +118,10 @@ namespace NIST.CVP.Generation.Core.Tests
         {
             GenerateResponse result = null;
             var fileNameRoot = Guid.NewGuid();
-
-            try
-            {
-                result = _subject.Generate(Path.Combine(_testPath, $"{fileNameRoot.ToString()}.json"));
-            }
-            finally
-            {
-                // Find and delete files as a result of the test
-                var files = Directory.GetFiles(_testPath, $"{fileNameRoot}*").ToList();
-                if (files.Count <= 4)
-                {
-                    foreach (var file in files)
-                    {
-                        File.Delete(file);
-                    }
-                }
-            }
+            result = _subject.Generate(new GenerateRequest(string.Empty));
 
             Assert.IsTrue(result.Success, result.ErrorMessage);
             Assert.AreEqual(StatusCode.Success, result.StatusCode);
-        }
-
-        [Test]
-        public void ShouldProperlySaveOutputsForEachResolverWithValidFiles()
-        {
-            var subject = new FakeGenerator(
-                _mockITestVectorFactory.Object, 
-                _mockIParameterParser.Object, 
-                _mockIParameterValidator.Object, 
-                _mockITestCaseGeneratorFactoryFactory.Object, 
-                _mockIVectorSetSerializer.Object
-            );
-            var testVectorSet = new FakeTestVectorSet();
-            var result = subject.SaveOutputsTester(_testPath, testVectorSet);
-            Assert.IsTrue(result.Success);
-        }
-
-        [Test]
-        public void ShouldNotSaveOutputsForEachResolverWithInvalidPath()
-        {
-            var subject = new FakeGenerator(
-                _mockITestVectorFactory.Object,
-                _mockIParameterParser.Object,
-                _mockIParameterValidator.Object,
-                _mockITestCaseGeneratorFactoryFactory.Object,
-                _mockIVectorSetSerializer.Object
-            );
-            var testVectorSet = new FakeTestVectorSet();
-            var jsonPath = Path.Combine(_testPath, "fakePath/");
-            var result = subject.SaveOutputsTester(jsonPath, testVectorSet);
-            Assert.IsFalse(result.Success);
-            Assert.AreEqual(StatusCode.FileSaveError, result.StatusCode);
         }
     }
 }
