@@ -1,5 +1,7 @@
 ﻿using System.IO;
+using Newtonsoft.Json;
 using NIST.CVP.Generation.Core.Parsers;
+using NIST.CVP.Generation.Core.Tests.Fakes;
 using NIST.CVP.Tests.Core;
 using NIST.CVP.Tests.Core.TestCategoryAttributes;
 using NUnit.Framework;
@@ -9,42 +11,28 @@ namespace NIST.CVP.Generation.Core.Tests.Parsers
     [TestFixture, UnitTest]
     public class ParameterParserTests
     {
-        private string _unitTestPath;
-
-        [OneTimeSetUp]
-        public void SetUp()
+        private ParameterParser<FakeParameters> _subject = new ParameterParser<FakeParameters>();
+        
+        public void ShouldParseParametersCorrectly()
         {
-            _unitTestPath = Utilities.GetConsistentTestingStartPath(GetType(), @"..\..\TestFiles");
-        }
+            var parameters = new FakeParameters()
+            {
+                Algorithm = "test test",
+                Mode = "also test",
+                Revision = "Special Publication",
+                IsSample = true,
+                VectorSetId = 42,
+            };
 
-        [Test]
-        public void ShouldReturnErrorForNonExistentPath()
-        {
-            var subject = new ParameterParser<IParameters>();
-            var result = subject.Parse(@"path/does/not/exist/file.json");
-            Assert.IsNotNull(result);
-            Assert.IsFalse(result.Success);
-        }
+            var parametersJson = JsonConvert.SerializeObject(parameters);
 
-        [Test]
-        [TestCase("")]
-        [TestCase(null)]
-        public void ShouldReturnErrorForNullOrEmptyPath(string path)
-        {
-            var subject = new ParameterParser<IParameters>();
-            var result = subject.Parse(path);
-            Assert.IsNotNull(result);
-            Assert.IsFalse(result.Success);
-        }
-
-        [Test]
-        public void ShouldReturnErrorForBadFile()
-        {
-            var subject = new ParameterParser<IParameters>();
-            var path = Path.Combine(_unitTestPath, "notJson.json");
-            var result = subject.Parse(path);
-            Assert.IsNotNull(result);
-            Assert.IsFalse(result.Success);
+            var result = _subject.Parse(parametersJson);
+            
+            Assert.AreEqual(parameters.Algorithm, result.ParsedObject.Algorithm, nameof(parameters.Algorithm));
+            Assert.AreEqual(parameters.Mode, result.ParsedObject.Mode, nameof(parameters.Mode));
+            Assert.AreEqual(parameters.Revision, result.ParsedObject.Revision, nameof(parameters.Revision));
+            Assert.AreEqual(parameters.IsSample, result.ParsedObject.IsSample, nameof(parameters.IsSample));
+            Assert.AreEqual(parameters.VectorSetId, result.ParsedObject.VectorSetId, nameof(parameters.VectorSetId));
         }
     }
 }
