@@ -7,19 +7,9 @@ namespace NIST.CVP.Generation.GenValApp.Tests.Fakes
 {
     public class FakeValidator : IValidator
     {
-        public ValidateResponse Validate(string resultPath, string answerPath, bool showExpected = false)
+        public ValidateResponse Validate(ValidateRequest validateRequest)
         {
-            if (resultPath.Contains("bad"))
-            {
-                return new ValidateResponse("fail");
-            }
-
-            if (answerPath.Contains("bad"))
-            {
-                return new ValidateResponse("fail");
-            }
-
-            return new ValidateResponse();
+            return new ValidateResponse(string.Empty);
         }
     }
 }
