@@ -5,13 +5,8 @@ namespace NIST.CVP.ParameterChecker.Tests.Fakes
 {
     public class FakeParameterChecker : IParameterChecker
     {
-        public ParameterCheckResponse CheckParameters(string requestFilePath)
+        public ParameterCheckResponse CheckParameters(ParameterCheckRequest request)
         {
-            if (requestFilePath.Contains("bad") || requestFilePath.Contains("not"))
-            {
-                return new ParameterCheckResponse("fail");
-            }
-
             return new ParameterCheckResponse();
         }
     }
