@@ -1,10 +1,14 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using System;
+using System.Linq;
+using System.Net;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using NIST.CVP.Common.Config;
-using NIST.CVP.Common.Enums;
 using NIST.CVP.Common.Helpers;
 using NIST.CVP.Common.Interfaces;
 using NIST.CVP.Orleans.Grains;
@@ -13,13 +17,8 @@ using NIST.CVP.Orleans.ServerHost.Models;
 using Orleans;
 using Orleans.Configuration;
 using Orleans.Hosting;
-using Orleans.Logging;
 using Orleans.Statistics;
-using System;
-using System.Linq;
-using System.Net;
-using System.Threading;
-using System.Threading.Tasks;
+using Environments = NIST.CVP.Common.Enums.Environments;
 
 namespace NIST.CVP.Orleans.ServerHost
 {
@@ -141,11 +140,6 @@ namespace NIST.CVP.Orleans.ServerHost
                 if (_orleansConfig.UseConsoleLogging)
                 {
                     logging.AddConsole();
-                }
-
-                if (_orleansConfig.UseFileLogging)
-                {
-                    logging.AddProvider(new FileLoggerProvider($"{DateTime.UtcNow:yyyy-MM-dd_Hmm}_{_environmentConfig.Name}.log"));
                 }
             });
         }
